+++ chia/wallet/util/puzzle_compression.py
@@ -4,6 +4,7 @@ import zlib
 from typing import List
 
 from chia.types.blockchain_format.program import Program
+from chia.wallet.cat_wallet.cat_utils import CAT_MOD
 from chia.wallet.nft_wallet.nft_puzzles import (
     NFT_METADATA_UPDATER,
     NFT_OWNERSHIP_LAYER,
@@ -12,7 +13,6 @@ from chia.wallet.nft_wallet.nft_puzzles import (
     SINGLETON_TOP_LAYER_MOD,
 )
 from chia.wallet.puzzles import p2_delegated_puzzle_or_hidden_puzzle as standard_puzzle
-from chia.wallet.puzzles.cat_loader import CAT_MOD
 from chia.wallet.puzzles.load_clvm import load_clvm_maybe_recompile
 
 # Need the legacy CAT mod for zlib backwards compatibility
